pyproject.toml
[tool.poetry]
name = "ventas-python-back"
version = "0.1.0"
description = ""
authors = ["Danilo Ramirez <danilo0613@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
